+++ react/components/EXPERIMENTAL_Table/Sections/Row.tsx
@@ -35,13 +35,17 @@ function Row(
   ref: Ref<HTMLTableRowElement>
 ) {
   const LIGHT_BLUE = '#DBE9FD'
-  const { rowHeight, density } = useMeasuresContext()
+  const { density } = useMeasuresContext()
   const { columns } = useDataContext()
   const { highlightOnHover, isRowActive, onRowClick } = useBodyContext()
-  const className = classNames('w-100 truncate overflow-x-hidden', {
-    'pointer hover-c-link': !header && onRowClick,
-    'hover-bg-muted-5': !header && (highlightOnHover || !!onRowClick),
-  })
+  const className = classNames(
+    'w-100 truncate overflow-x-hidden',
+    {
+      'pointer hover-c-link': !header && onRowClick,
+      'hover-bg-muted-5': !header && (highlightOnHover || !!onRowClick),
+    },
+    props.className
+  )
   const rowColor =
     data && isRowActive && isRowActive(data)
       ? { backgroundColor: LIGHT_BLUE }
@@ -88,13 +92,13 @@ function Row(
         const content = cellRenderer
           ? cellRenderer({
               data: cellData,
-              rowHeight,
+              rowHeight: height,
               density,
               motion,
             })
           : cellData
         return (
-          <Cell key={id} width={width}>
+          <Cell key={id} width={width} height={height}>
             {content}
           </Cell>
         )