+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.117.0] - 2020-05-13
+
+### Added
+
+- `hideColumns` & `showColumns` functions to `useTableVisibility`
+- `bodyHeight` memo to `useTableMeasures`
+- `disableScroll` to `Table.Sections`
+
+### Fixed
+
+- `Table.Body.Cell` unstable heights
+
 ## [9.116.0] - 2020-05-07
 
 ### Added